# apps/api/docs/swagger.yaml
definitions:
  github_com_projectx_api_internal_platform_httpx.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  internal_reviews.CreateReviewInput:
    properties:
      name:
        maxLength: 255
        minLength: 2
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      review_text:
        maxLength: 5000
        minLength: 10
        type: string
    required:
    - name
    - rating
    - review_text
    type: object
  internal_reviews.Review:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      rating:
        type: integer
      review_text:
        type: string
    type: object
...
info:
  contact:
    name: Project-X support
  description: |-
    REST API for Project-X — the AI-run marketing + studio operations platform.
    Auth is a JWT stored in an HTTP-only cookie (set by POST /api/v1/auth/login).
    When calling protected endpoints from a tool like Swagger UI or curl, log in first and reuse the returned cookie.
  license:
    name: Proprietary
  title: Project-X API
  version: "1.0"
paths:
...
  /api/v1/me/studios/{id}/billing/portal:
    post:
      description: Looks up the studio's Stripe customer by contact email against
        the platform's Stripe account and creates a Stripe billing-portal session
        so the studio can manage their existing subscription (update payment method,
        view invoices, cancel, etc.). Returns the portal session URL to redirect the
        studio to.
      parameters:
      - description: Studio ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: global studio, invalid studio ID, or no active subscription
            to manage
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "404":
          description: studio not found
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: platform Stripe account not configured, or Stripe error creating
            the session
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create a Stripe billing portal session
      tags:
      - Billing
  /api/v1/me/studios/{id}/billing/sync:
    post:
      description: Checks Stripe directly for the studio's subscription status and
        updates the DB accordingly. This is the fallback for when webhooks don't fire
        (e.g. local dev). Looks up the Stripe customer by contact email, finds an
        active/trialing subscription, resolves the plan tier from subscription metadata
        (falling back to matching checkout session metadata), and updates the studio's
        stored subscription tier to canceled, past_due-restored, or the resolved tier
        as appropriate. In the local environment (API_ENV=local) it reports status
        without writing any DB changes.
      parameters:
      - description: Studio ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: global studio or invalid studio ID
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "404":
          description: studio not found
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Sync billing status from Stripe
      tags:
      - Billing
...
  /api/v1/reviews:
    get:
      description: Public endpoint returning every submitted review. No auth required.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_reviews.Review'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      summary: List all reviews
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Public endpoint for a visitor to submit a testimonial/review. No
        auth required.
      parameters:
      - description: Review payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_reviews.CreateReviewInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_reviews.Review'
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      summary: Submit a review
      tags:
      - Reviews
...
  /api/v1/webhooks/x:
    get:
      consumes:
      - application/json
      description: Handles both halves of X's Account Activity webhook. On GET, responds
        to X's CRC challenge by HMAC-SHA256-signing the `crc_token` query param with
        the connected channel's consumer secret and returning the base64-encoded `response_token`.
        On POST, decodes the `direct_message_events` payload and dispatches each inbound
        (non-self) message_create event to the messaging service, always returning
        200. No session/API-key auth — GET is authenticated implicitly by the shared
        consumer secret used to answer the CRC challenge; POST performs no signature
        verification.
      parameters:
      - description: X CRC challenge token (GET only)
        in: query
        name: crc_token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: missing crc_token or no active X channel found (GET only)
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      summary: X (Twitter) webhook CRC challenge and inbound DM events
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Handles both halves of X's Account Activity webhook. On GET, responds
        to X's CRC challenge by HMAC-SHA256-signing the `crc_token` query param with
        the connected channel's consumer secret and returning the base64-encoded `response_token`.
        On POST, decodes the `direct_message_events` payload and dispatches each inbound
        (non-self) message_create event to the messaging service, always returning
        200. No session/API-key auth — GET is authenticated implicitly by the shared
        consumer secret used to answer the CRC challenge; POST performs no signature
        verification.
      parameters:
      - description: X CRC challenge token (GET only)
        in: query
        name: crc_token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: missing crc_token or no active X channel found (GET only)
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      summary: X (Twitter) webhook CRC challenge and inbound DM events
      tags:
      - Webhooks
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    description: JWT session cookie set by POST /api/v1/auth/login (cookie name configurable
      via COOKIE_NAME).
    in: cookie
    name: px_session
    type: apiKey
swagger: "2.0"